MAJOR REFACTOR: Fixing and breaking things
I have fixed the following:
-almost nothing

What I broke:
- a lot

What's going to happen next?

- After lots of experimenting it seems the best methodology to continue is to go to a one camera/target system, and figure out how to determine the distance between an object and the player.

If the object occludes the player and is close to the camera, clip! If the camera is a fair bit behind the object, don't clip.

This is going to be a pain and will probably introduce a few new problems, namely...

- If there's more than one object between the player, and the camera does decide to clip, it will clip staight to the player.  Have to do research to see if this is normal camera behavior in 3D games, or if it will clip down to the next nearest object.