BUG FIX: Collision turning off while in contact with multiplie bodies
Discovered a bug where it was possible that the camera would stop clipping and detecting collisions when the CollisionProbe left a body while in contact with another body.

Solution: add array to register each body, and remove it when the body left the CollisionProbe. Only turn off collision when this array is empty, and is_clipping is false.  Bug has been squashed, but we'll see what new ones will pop up from this solution...