FIXED: Camera/Player Orientations
Fixed the orientations so that the camera and player are both facing -Z, which is forwards in Godot for some reason.

Broke rotations code until I fixed it by adding - signs in front of the variables to flip them properly...

Going to attempt to add a mode where the camera looks at the player and will rotate to keep them in focus and only move when the player is far enough away, but due to nodes and transforms this may not be possible in my current setup.

May need some black magic to get everything working properly.