ADDED: Player Orbiting Camera
This was hard.

Godot seems to do weird things with rotations and it really confused me.

I finally added the ability to have the player "orbit" the camera, which allows the player to go left or right and have the camera look at them, and it will eventually lead to the camera following them.

Need to refactor a metric shitton of code here soon, it's getting ridiculous. Lots of unused variables from failed experience, but I *always* push commits when I finally get a feature working! And make really long-winded messages...

This commit doesn't change much other than some basic rotation code that already existed in previous commits.

Oh, and I also replaced the Area that determines if the player is in a confined space or not with a raycast, since that is a more accurate way of determing that state on the fly. Areas can get bugged out easily sometimes.