FIXED: Better Player Orbiting
This is much closer to what I want for a "player orbiting camera" feature, similar to what you would see in games like Super Mario 64, and most other typical 3D platformers.

Although the camera motion is still a bit rigid, this is *really* close to what I want for my game, so, I'm comitting it.