Tidied Up some Camera Movement
Tidied up the camera's movement by making the camera lerp at anytime other than when its clipping - and also ensured that if the camera is clipping to geometry BEHIND the player after clipping to something closer to the player, it will lerp back slower to that collision point, instead of just instantly snapping backwards. Really improves the overall feel of the camera and results in much nicer, smoother feeling gameplay.